# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.Cabin:
    properties:
      bed_type:
        type: string
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
    type: object
  dto.CabinDetailResponse:
    properties:
      bed_type:
        type: string
      capacity:
        type: integer
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  dto.Cabins:
    properties:
      available_cabins:
        type: integer
      occupied_units:
        items:
          type: string
        type: array
      total_cabins:
        type: integer
    type: object
...
  dto.CreateCabinRequest:
    properties:
      cabins:
        items:
          $ref: '#/definitions/dto.Cabin'
        type: array
    type: object
...
  dto.CreateSeatRequest:
    properties:
      is_active:
        type: boolean
      left_columns_count:
        type: integer
      nivel:
        type: integer
      prefix:
        type: boolean
      right_columns_count:
        type: integer
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: detailed error message
        type: string
    type: object
...
  dto.ListUnitAvailableResponse:
    properties:
      cabins:
        $ref: '#/definitions/dto.Cabins'
      seats:
        $ref: '#/definitions/dto.Seats'
      ship_id:
        type: string
      ship_name:
        type: string
    type: object
...
  dto.SeatDetailResponse:
    properties:
      id:
        type: string
      is_active:
        type: boolean
      left_columns_count:
        type: integer
      prefix:
        type: boolean
      right_columns_count:
        type: integer
    type: object
  dto.Seats:
    properties:
      available_seats:
        type: integer
      occupied_units:
        items:
          type: string
        type: array
      total_seats:
        type: integer
    type: object
...
  dto.UpdateCabinRequest:
    properties:
      bed_type:
        type: string
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdateSeatRequest:
    properties:
      is_active:
        type: boolean
      left_columns_count:
        type: integer
      prefix:
        type: boolean
      right_columns_count:
        type: integer
    type: object
...
host: localhost:8002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Service for managing user identities, roles, and organizations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Uber dos Rios API
  version: v1.0.16
paths:
...
  /ships-config/cabins:
    get:
      consumes:
      - application/json
      description: Retrieves a cabins.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CabinDetailResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List Cabins
      tags:
      - cabins
    post:
      consumes:
      - application/json
      description: Registers a new cabin with the provided details.
      parameters:
      - description: Cabin Data to Create
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCabinRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CabinDetailResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new cabin
      tags:
      - cabins
  /ships-config/cabins/{id}:
    delete:
      description: Delete a cabin by ID
      parameters:
      - description: Cabin ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID supplied
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Seat not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete cabin
      tags:
      - cabins
    put:
      consumes:
      - application/json
      description: Updates an existing cabin with the provided details.
      parameters:
      - description: Cabin Data to Update
        in: body
        name: cabin
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCabinRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SeatDetailResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a cabin
      tags:
      - cabins
  /ships-config/seats:
    get:
      consumes:
      - application/json
      description: Retrieves a seats.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SeatDetailResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List seats
      tags:
      - seats
    post:
      consumes:
      - application/json
      description: Registers a new seat with the provided details.
      parameters:
      - description: Seat Data to Create
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSeatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SeatDetailResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new seat
      tags:
      - seats
  /ships-config/seats/{id}:
    delete:
      description: Delete a seat by ID
      parameters:
      - description: Seat ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID supplied
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Seat not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete seat
      tags:
      - seats
    put:
      consumes:
      - application/json
      description: Updates an existing seat with the provided details.
      parameters:
      - description: Seat Data to Update
        in: body
        name: seat
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSeatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SeatDetailResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a seat
      tags:
      - seats
  /ships-config/units/search:
    get:
      consumes:
      - application/json
      description: Searches for seats based on available date range.
      parameters:
      - description: Data de Início (YYYY-MM-DD)
        in: query
        name: dateorigin
        required: true
        type: string
      - description: Data de Fim (YYYY-MM-DD)
        in: query
        name: datedest
        required: true
        type: string
      - description: 'ID Barco: uuid'
        in: query
        name: shipid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ListUnitAvailableResponse'
            type: array
      summary: Search Available Seats
      tags:
      - seats
...
schemes:
- http
- https
swagger: "2.0"